Restructure code somewhat
I’m not fond of some of these names.

I’m not fond of deep paths.

Yet I think that overall this restructuring of access is an improvement.
A little.
Maybe it needs to go further.

The trouble starts with tablet-v2’s own names for things.
Everything is zwp_tablet_*_v2… except zwp_tablet_v2.
If you viewed “tablet” as the namespace,
it’d mean you have types Manager, Seat, Tool, Pad, PadGroup, PadRing and PadStrip…
and a type that’s not called Tablet because it actually just has no name.
A part of me wishes it was actually named zwp_tablet_tablet_v2.

It gets worse when you have this notion of a “tablet seat”.

At present I have modules named sctk::seat::tablet::seat and sctk::seat::tablet::tablet.
I find this rather amusing, but freely admit it to be suboptimal.

Maybe we need a different convention.
As a general rule in Rust, you import types and traits individually,
functions by module, constants vary,
and you mostly avoid duplication of the module name in type names.
Across a large code base you may get many things called “Error”,
and if you want to deal with more than one, you rename the import.
But there are some places where a different convention holds:
you don’t import each thing from std::fmt, but just import the module,
and `impl fmt::Display`, referring to fmt::Formatter and fmt::Result.

If I controlled *all* the naming, I’d probably structure things so:

    sctk::seat::{
        TabletManager (mismatching sctk’s *State custom,
                       but matching zwp_tablet_manager_v2,
                       and notably not exposed in a submodule),
        tablet_seat::{Data (hidden), Handler},
        tablet::{
            Data (hidden),
            Description (accumulator for the initial burst of events),
            Event,
            Handler,
        },
        tablet_tool::{
            Capability (reexported enum),
            Data (hidden),
            Description (accumulator for the initial burst of events),
            Event,
            EventList (maybe gone?),
            Frame,
            Handler,
            InitEvent,
            InitEventList (maybe gone?),
            State (optional accumulator of frames,
                   nothing to do with sctk’s *State convention),
            Type (reexported enum),
        },
        tablet_pad::{
            I dunno, not so comfortable with the further namespacing here,
            we have pad, pad group, pad ring and pad strip.
            Might even rename the module just pad.
        },
    }

And users would generally be expected to use it like this:

    use sctk::seat::{TabletManager, tablet_seat, tablet, tablet_tool, tablet_pad};
    impl tablet_tool::Handler for … {
        fn …(
            …,
            events: Vec<tablet_tool::Event>,
        ) {
            if self.tools.get(tablet).type == tablet_tool::Type::Pen {

Another possibility would be putting everything inside sctk::seat::tablet:

    sctk::seat::tablet::{
        Manager, (spelling: tablet::Manager)
        seat::{Data, Handler}, (spelling: tablet::seat::Handler)
        tool::{…}, (spelling: tablet::tool::Handler)

        // Tablet itself is messy.
        // One possibility (spelling: tablet::tablet::Handler):
        tablet::{…},
        // Another possibility (spelling: tablet::Event):
        // Not so fond of this, especially if it led to tablet::State!
        …,

        // One possibility:
        pad::{…},
        pad_group::{…},
        pad_ring::{…},
        pad_strip::{…},
        // Another possibility:
        pad::{
            …,
            group::{…},
            ring::{…},
            strip::{…},
        },
    }

But tablet::tool::Handler is feeling a little icky,
and then we’re back with sctk::seat::tablet::seat.

There are at least three sets of customs: general Rust, Wayland, sctk.
It’s not possible to satisfy all three,
and at places it may be a struggle to satisfy even two.
I think we’re going to be breaking someone’s custom, somewhere,
however we do it.
